# docs/02-openapi-spec.yaml
openapi: 3.0.3
info:
  title: TikTok Shop AI Video System API
  version: 1.0.0
  description: |
    Java Spring Boot 主业务后端 API。
    前端只调用本 API。
    Python AI 服务和 Render Worker 通过 callback 接口回调本 API。
    V1 契约版本为 1.0.0。AI 回调、RenderManifest 和数据库记录必须保存对应版本号。
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/video-tasks:
    post:
      tags: [VideoTasks]
      summary: Create video generation task
      description: |
        创建视频任务时预扣 1 次成片额度。
        Java 创建 task 后调用 Python AI Orchestrator 启动第一阶段工作流。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoTaskRequest"
      responses:
        "200":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateVideoTaskResponse"

    get:
      tags: [VideoTasks]
      summary: List current user's video tasks
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/VideoTaskStatus"
        - name: productId
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskListResponse"
...
  /api/video-tasks/{taskId}:
    get:
      tags: [VideoTasks]
      summary: Get task status and detail
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskDetailResponse"
...
  /api/video-tasks/{taskId}/select-plan:
    post:
      tags: [VideoTasks]
      summary: Select a video plan
      description: |
        用户选择一个视频方案后，Java 保存 selectedPlanId，
        然后调用 Python AI 服务启动第二阶段脚本、分镜、素材生成工作流。
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties:
                planId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Plan selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
  /api/video-tasks/{taskId}/retry:
    post:
      tags: [VideoTasks]
      summary: Retry failed task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Retry accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
  /api/video-tasks/{taskId}/cancel:
    post:
      tags: [VideoTasks]
      summary: Cancel an in-progress task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"

  # ── Fashion Creative Loop V1 (M2) ──
...
  /api/video-tasks/{taskId}/confirm-plan:
    post:
      tags: [VideoTasks]
      summary: Confirm video plan and proceed to storyboard generation
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPlanRequest"
      responses:
        "200":
          description: Plan confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
  /api/video-tasks/{taskId}/confirm-storyboard:
    post:
      tags: [VideoTasks]
      summary: Confirm storyboard and proceed to keyframe configuration
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Storyboard confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
...
  /api/video-tasks/{taskId}/render:
    post:
      tags: [VideoTasks]
      summary: Request video render from confirmed clips
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Render requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
  /api/video-tasks/{taskId}/approve:
    post:
      tags: [VideoTasks]
      summary: Approve final video review
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Final video approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
  /api/video-tasks/{taskId}/feedback:
    post:
      tags: [VideoTasks]
      summary: Submit feedback for repair
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairEventListResponse"
  /api/video-tasks/{taskId}/repair-events:
    get:
      tags: [VideoTasks]
      summary: List repair events for a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Repair event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairEventListResponse"
...
  /api/videos/{videoId}:
    get:
      tags: [Videos]
      summary: Get video detail
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: Video detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetailResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VideoId:
      name: videoId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    VideoType:
      type: string
      enum:
        - pain_point_solution
        - before_after
        - review
        - product_showcase
        - ugc_style
        - tutorial
    VideoTaskStatus:
      type: string
      enum:
        - draft
        - asset_uploading
        - asset_analyzing
        - waiting_asset_confirmation
        - reference_analyzing
        - plan_generating
        - analyzing
        - analysis_completed
        - plan_generated
        - waiting_plan_selection
        - storyboard_generating
        - script_generating
        - script_generated
        - material_generating
        - material_generated
        - rendering
        - checking
        - completed
        - failed
        - exported
        - waiting_storyboard_confirmation
        - keyframe_configuring
        - image_generating
        - waiting_image_confirmation
        - video_clip_generating
        - waiting_video_clip_confirmation
        - waiting_final_review
        - repairing
        - cancelled
    TaskMode:
      type: string
      enum:
        - PRODUCT_CREATIVE
        - REFERENCE_STORYBOARD
        - USER_SCRIPT
        - CUSTOM_STORYBOARD
    VideoStatus:
      type: string
      enum:
        - completed
        - exported
        - deleted
...
    CreateVideoTaskRequest:
      type: object
      required: [productId, duration]
      properties:
        productId:
          type: string
          format: uuid
        duration:
          type: integer
          enum: [15, 20, 25, 30]
        videoType:
          $ref: "#/components/schemas/VideoType"
          description: Legacy video structure hint. Optional for Fashion Creative Loop modes; AI creative_plan will produce the actual structure candidates.
        needSubtitles:
          type: boolean
          default: true
        needVoiceover:
          type: boolean
          default: false
        taskMode:
          $ref: "#/components/schemas/TaskMode"
        productCategory:
          type: string
          nullable: true
        shotCount:
          type: integer
          nullable: true
...
    CreateVideoTaskResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/VideoTaskStatus"
            progress:
              type: integer
    VideoTask:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/VideoTaskStatus"
        progress:
          type: integer
        duration:
          type: integer
        videoType:
          $ref: "#/components/schemas/VideoType"
        needSubtitles:
          type: boolean
        needVoiceover:
          type: boolean
        selectedPlanId:
          type: string
          format: uuid
          nullable: true
        selectedPlanTitle:
          type: string
          nullable: true
        assetAnalysis:
          allOf:
            - $ref: "#/components/schemas/FashionAssetAnalysis"
          nullable: true
        taskMode:
          $ref: "#/components/schemas/TaskMode"
        productCategory:
          type: string
          nullable: true
        shotCount:
          type: integer
          nullable: true
        currentVersion:
          type: integer
        errorMessage:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
        failedStage:
          type: string
          nullable: true
        errorRetryable:
          type: boolean
          nullable: true
        manifestVersion:
          type: string
          example: 1.0.0
        schemaVersion:
          type: string
          example: 1.0.0
        retryCount:
          type: integer
    FashionAssetAnalysis:
      type: object
      required: [schemaVersion, analysisText, analyzedAssetIds, model, analyzedAt]
      properties:
        schemaVersion:
          type: string
          enum: ["1.0"]
        analysisText:
          type: string
        analyzedAssetIds:
          type: array
          items:
            type: string
        model:
          type: string
        analyzedAt:
          type: string
          format: date-time
    VideoTaskDetailResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/VideoTask"
    VideoTaskStatusResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/VideoTaskStatus"
            progress:
              type: integer
    VideoTaskListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/VideoTask"
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
...
    Video:
      type: object
      properties:
        videoId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        title:
          type: string
        videoUrl:
          type: string
        coverUrl:
          type: string
        duration:
          type: integer
        resolution:
          type: string
        qualityScore:
          type: integer
        riskScore:
          type: integer
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/VideoStatus"
    VideoDetailResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/Video"
...
    RepairEvent:
      type: object
      properties:
        repairEventId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        targetType:
          type: string
          enum: [storyboard, keyframe, video_clip, render_manifest, final_video, plan]
        targetId:
          type: string
          nullable: true
        userFeedback:
          type: string
        issueType:
          type: string
          nullable: true
        repairScope:
          type: object
          nullable: true
        repairPlan:
          type: object
          nullable: true
        beforeVersion:
          type: integer
          nullable: true
        afterVersion:
          type: integer
          nullable: true
        status:
          type: string
          enum: [created, in_progress, completed, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RepairEventListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
            events:
              type: array
              items:
                $ref: "#/components/schemas/RepairEvent"
    FeedbackRequest:
      type: object
      required: [feedbackText]
      properties:
        feedbackText:
          type: string
        category:
          type: string
        targetType:
          type: string
          enum: [storyboard, keyframe, video_clip, render_manifest, final_video, plan]
        targetId:
          type: string
          nullable: true
...
    ConfirmPlanRequest:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
          format: uuid
...